+++ examples/complete/data.tf
@@ -4,7 +4,7 @@ data "aws_ami" "amazon_linux" {
 
   filter {
     name   = "name"
-    values = ["amzn-ami-hvm-*-x86_64-gp2"]
+    values = ["amzn2-ami-*-x86_64-gp2"]
   }
 }
 
@@ -12,12 +12,6 @@ data "aws_availability_zones" "available" {
   state = "available"
 }
 
-data "aws_vpc" "default" {
-  default = true
-}
+data "aws_caller_identity" "current" {}
 
-data "aws_subnet" "default" {
-  vpc_id            = data.aws_vpc.default.id
-  availability_zone = data.aws_availability_zones.available.names[0]
-  default_for_az    = true
-}
+data "aws_region" "current" {}